backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyjwt==2.11.0
bcrypt==4.1.1
passlib==1.7.4
openai==1.12.0
pdfplumber==0.10.3
python-docx==0.8.11
aiofiles==23.2.1
pytest==8.3.4
